Fixed bug in GMRES solver
The call to the nextIteration() method must occur exactly once per
iteration, otherwise incrementing the total iterations counter in
wrong places might cause serious bugs. For example if minIterations is
equal to 1 and the initial residue is smaller than convergenceResidue,
the loop on line 154 in tnlGMRESSolver_impl.h will be entered, but the
m-loop on line 179 will be skipped and on line 263 update() will be
called with empty vectors, which will lead to division by zero.