Removed check for HAVE_CUSPARSE
It is not needed, since CUSPARSE is included in all versions of CUDA
since at least 7.0.

Also note that CUSPARSE is needed only for some tests and legacy code
(tnl-benchmark-spmv and wrappers in Matrices::CSR) and experimental code
(Preconditioners::ILU0). Everything is still guarded by the
HAVE_CUSPARSE macro, so the user can pass -DHAVE_CUSPARSE -lcusparse to
the compiler to enable the features.