Removed the remaining autotools-like macros HAVE_*_H from tnlConfig.h
TNL is not supposed to have any external dependencies, so an empty
config file is useless. If some marginal parts of the project (such as
code for image processing) need to depend on some non-standard library,
they should be split into a separate module/project so that the
dependencies can be properly specified.

For now, the macros are specified as compiler parameters. Client
software using the affected headers will have to do the same.