Moved everything from Cuda.cpp and Cuda.cu into Cuda_impl.h
Having 4 different files (Cuda.h, Cuda_impl.h, Cuda.cpp, Cuda.cu) for
one class is not clear at all. On a related note, a class with 28
methods is clearly too long to be considered as well-designed, so it
should be split into several subclasses. Obviously, the full CUDA
support cannot be provided/wrapped by a single class so the design
should be more flexible.