distributeSubentities: fix the algorithm for assigning global IDs
Previously we matched the local subvertex ID with the neighbor's local
superentity ID (entities sharing the same vertex), which does not make
sense. The result cannot use the synchronizeSparse function, so it looks
more complicated.

Note that it still does not work in all corner cases - see the FIXME
notes.