Mesh writers: avoid reference in class members, initialize by rdbuf
Using references as class members is weird, because they cannot extend
the lifetime of the objects they are initialized with. Using values of
the type std::ostream and initializing by rdbuf (which is a pointer)
works better, probably because the underlying rdbuf generally outlives
the ostreams that were being passed to the writers.