Removed the code for entity orientations in the unstructured mesh
It was unused, untested and unnecessarily complicated. We will deal with
entity orientations when it is actually needed.

I think that the implementation based on storing the whole permutations
of subentity indices is not appropriate anyway - it induces huge memory
consumption, while e.g. for faces we are interested only in the
orientation of the outward normal vector with respect to the two
adjacent cells, which can be stored as one bit per face-cell pair, or it
can be deduced from the entity indices if we enforce a suitable
convention (e.g. all normal vectors could be outward with respect to the
adjacent cell with the smallest index). I think OpenFOAM uses a similar
convention for normal vectors. As for the orientation of edges in 3D, I
have no idea where it could be useful...