Replaced copyright headers in files under src/TNL/ with a simple and maintainable template
The old headers were hard to maintain, inconsistent and contained
useless information:

- file name (useless, prone to inconsistencies upon renaming)
- "description" (useless placeholder that was likely intended to be
  replaced by the user, not to be used literally)
- "begin" date for each file (prone to inconsistencies and copy-paste
  errors, the information can be found in git)
- "copyright" notice (inconsistent years)
- "email" (replaced with a project link in the new template)
- "Copyright Notice" referring to an obsolete file (tnl/Copyright) that
  does not exist anymore

The new template consists of three simple parts:

- copyright notice with updatable years related to the whole project
- project name and website URL
- SPDX license identifier

Note that the SPDX license identifier must be on a line that starts with
either "//" or "/*", so we can't use a block comment (/* ... */) for the
whole header. But using "//" for block comments is common too.